requirements.txt
arrow==0.10.0
Flask==0.12.1
Flask-Login==0.4.0
flask-mongoengine==0.9.2
Flask-WTF==0.14.2
itsdangerous==0.24
praw==4.5.1
